Moved save_audio to a utils folder and refactored code accordingly.
