+++ pyneuphonic/__init__.py
@@ -1,2 +1,3 @@
 from pyneuphonic.client import Neuphonic
 from pyneuphonic.models import TTSConfig, WebsocketEvents, AgentConfig
+from pyneuphonic._utils import save_audio
